refactor(timeline): split impl RoomScreen by concern (final stage)
Stage 5 of the room_screen split, verbatim code motion: the giant
impl RoomScreen breaks into per-concern impl blocks living next to
their domains — the timeline-update processing loop and its timers
(updates.rs), message-action / link / media / translation handlers
(interactions.rs), message search + jump-to-event (search.rs), and the
bot-admin, threads-pane, and room-info methods appended to their
existing domain files. Shared helpers stay in mod.rs; single-consumer
helpers and constants move with their only caller.

mod.rs: 5,628 -> 2,633 lines, completing the split of the original
17,790-line room_screen.rs into 15 focused files (largest 2.6k).
Verified as every prior stage: clean build, 582/582 lib tests, the
same 112 #[test] fns, byte-identical moved lines, and a clean app boot
with zero template resolution errors.
